created admin homepage
Homepage is responsive, collapses the sidebar for smaller screens and replaces it with a hamburger menu through css grid and BreakpointObserver. Currently no header present for the page and names and anchors of links are placeholders. Remaining pages for admin are : Members sh'd show alumni members, events should show and allow for creation of events, announce should allow for sending of bulk messages ,reports should be for generating reports from alumni db, also pages for changing website graphics and optionally changing of admin credentials